Continuing to trim down unused code in the modified ghash.
There was a bug in the caching logic, where an existing hash table would
have it's metadata read from NVM, but the cache would be initialized to
all 0's without loading from the device. This was causing files to be
reallocated on subsequent tests, rather that being read-only. Fixed by
filling the cache in g_hash_table_new and adding a FORCE flag to the
nvram_read function.